Updates for code review
