+++ ankimorphs/__init__.py
@@ -53,7 +53,7 @@ from .ankimorphs_db import AnkiMorphsDB
 from .extra_settings import ankimorphs_extra_settings, extra_settings_keys
 from .extra_settings.ankimorphs_extra_settings import AnkiMorphsExtraSettings
 from .generators.generators_window import GeneratorWindow
-from .jit_highlight_morphs import am_highlight_morphs
+from .highlight_morphs_jit import highlight_morphs_jit
 from .known_morphs_exporter import KnownMorphsExporterDialog
 from .progression.progression_window import ProgressionWindow
 from .recalc import recalc_main
@@ -99,7 +99,7 @@ def main() -> None:
 
     gui_hooks.profile_will_close.append(cleanup_profile_session)
 
-    hooks.field_filter.append(am_highlight_morphs)
+    hooks.field_filter.append(highlight_morphs_jit)
 
 
 def init_toolbar_items(links: list[str], toolbar: Toolbar) -> None:
